+++ examples/eqPixelBench/channel.cpp
@@ -203,9 +203,8 @@ void Channel::_testFormats( float applyZoom )
             }
 
             // write
-            eq::Compositor::ImageOp op;
+            eq::ImageOp op;
             op.image = image;
-            op.channel = this;
             op.buffers = eq::Frame::BUFFER_COLOR;
             op.offset = offset;
             op.zoom = zoom;
@@ -215,7 +214,7 @@ void Channel::_testFormats( float applyZoom )
             try
             {
                 clock.reset();
-                eq::Compositor::assembleImage( op );
+                eq::Compositor::assembleImage( op, this );
                 glFinish();
                 const float msec = clock.getTimef() / float( nLoops );
                 const GLenum error = glGetError(); // release mode
@@ -331,8 +330,7 @@ void Channel::_testTiledOperations()
         //---- benchmark assembly operations
         subPVP.y = pvp.y + tiles * subPVP.h;
 
-        eq::Compositor::ImageOp op;
-        op.channel = this;
+        eq::ImageOp op;
         op.buffers = eq::Frame::BUFFER_COLOR | eq::Frame::BUFFER_DEPTH;
         op.offset  = offset;
 
@@ -344,7 +342,7 @@ void Channel::_testTiledOperations()
         for( unsigned j = 0; j <= tiles; ++j )
         {
             op.image = images[j];
-            eq::Compositor::assembleImage( op );
+            eq::Compositor::assembleImage( op, this );
         }
         msec = clock.getTimef();
         _sendEvent( ASSEMBLE, msec, area, formatType.str(), 0, 0 );
@@ -421,8 +419,7 @@ void Channel::_testDepthAssemble()
                               "depthAssemble" );
 
         // benchmark
-        eq::Compositor::ImageOp op;
-        op.channel = this;
+        eq::ImageOp op;
         op.buffers = eq::Frame::BUFFER_COLOR | eq::Frame::BUFFER_DEPTH;
         op.offset  = offset;
 
@@ -434,7 +431,7 @@ void Channel::_testDepthAssemble()
         for( unsigned j = 0; j <= i; ++j )
         {
             op.image = images[j];
-            eq::Compositor::assembleImageDB_FF( op );
+            eq::Compositor::assembleImageDB_FF( op, this );
         }
         float msec = clock.getTimef();
         _sendEvent( ASSEMBLE, msec, area, formatType.str(), 0, 0 );
@@ -449,7 +446,7 @@ void Channel::_testDepthAssemble()
             for( unsigned j = 0; j <= i; ++j )
             {
                 op.image = images[j];
-                eq::Compositor::assembleImageDB_GLSL( op );
+                eq::Compositor::assembleImageDB_GLSL( op, this );
             }
             msec = clock.getTimef();
             _sendEvent( ASSEMBLE, msec, area, formatType.str(), 0, 0 );