+++ doc/Changelog.md
@@ -4,6 +4,9 @@ Changelog {#Changelog}
 # git master
 
 
+* [538](https://github.com/Eyescale/Equalizer/pull/538):
+  Support sorted image-based compositing. Minor API changes in the
+  compositor, frame and image code
 * [532](https://github.com/Eyescale/Equalizer/pull/532):
   Compositor::blendFrames() replaces Compositor::assembleFramesSorted()
 * [529](https://github.com/Eyescale/Equalizer/pull/529):